requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
numpy==1.26.4
pydantic==2.8.2
groq==0.31.1
python-dotenv==1.1.1
